backend/requirements/base.txt
Django==6.0.*
djangorestframework==3.15.*
drf-spectacular==0.27.*
drf-spectacular[sidecar]==0.27.*
django-environ==0.11.*
django-cors-headers==4.3.*
psycopg[binary]==3.1.*
PyJWT==2.8.*
cryptography==42.*
gunicorn==22.*
whitenoise==6.*
anthropic>=0.25,<1.0
openai>=1.30,<2.0
google-generativeai>=0.7,<1.0
django-redis>=5.4,<6.0
redis>=5.0,<6.0
pydantic>=2.7,<3.0
celery>=5.4,<6.0
stripe>=8.0.0
django-ratelimit>=4.1,<5.0
